backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.3
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
openpyxl==3.1.2
reportlab==4.1.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
